study_assistant/backend/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.28
asyncpg==0.29.0
pgvector==0.2.5
pydantic==2.6.4
pydantic-settings==2.2.1
alembic==1.13.1
google-genai==0.2.2
python-dotenv==1.0.1
